# Conflicts:
#	src/plugins/discover/public/application/angular/discover.js
#	x-pack/plugins/data_enhanced/public/search/search_interceptor.ts

+++ x-pack/plugins/data_enhanced/public/search/search_interceptor.ts
@@ -4,24 +4,17 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import { throwError, from, Subscription } from 'rxjs';
-import { tap, takeUntil, finalize, catchError } from 'rxjs/operators';
+import { throwError, Subscription } from 'rxjs';
+import { tap, finalize, catchError } from 'rxjs/operators';
 import {
   TimeoutErrorMode,
-  IEsSearchResponse,
   SearchInterceptor,
   SearchInterceptorDeps,
   UI_SETTINGS,
+  IKibanaSearchRequest,
 } from '../../../../../src/plugins/data/public';
-import { AbortError, abortSignalToPromise } from '../../../../../src/plugins/kibana_utils/public';
-
-import {
-  IAsyncSearchRequest,
-  ENHANCED_ES_SEARCH_STRATEGY,
-  IAsyncSearchOptions,
-  doPartialSearch,
-  throwOnEsError,
-} from '../../common';
+import { AbortError } from '../../../../../src/plugins/kibana_utils/common';
+import { ENHANCED_ES_SEARCH_STRATEGY, IAsyncSearchOptions, pollSearch } from '../../common';
 
 export class EnhancedSearchInterceptor extends SearchInterceptor {
   private uiSettingsSub: Subscription;
@@ -60,54 +53,30 @@ export class EnhancedSearchInterceptor extends SearchInterceptor {
     if (this.deps.usageCollector) this.deps.usageCollector.trackQueriesCancelled();
   };
 
-  public search(
-    request: IAsyncSearchRequest,
-    { pollInterval = 1000, ...options }: IAsyncSearchOptions = {}
-  ) {
-    let { id } = request;
-
+  public search({ id, ...request }: IKibanaSearchRequest, options: IAsyncSearchOptions = {}) {
     const { combinedSignal, timeoutSignal, cleanup, abort } = this.setupAbortSignal({
       abortSignal: options.abortSignal,
       timeout: this.searchTimeout,
     });
-    const abortedPromise = abortSignalToPromise(combinedSignal);
     const strategy = options?.strategy ?? ENHANCED_ES_SEARCH_STRATEGY;
+    const searchOptions = { ...options, strategy, abortSignal: combinedSignal };
+    const search = () => this.runSearch({ id, ...request }, searchOptions);
 
     this.pendingCount$.next(this.pendingCount$.getValue() + 1);
-
     const isCurrentSession =
       options.sessionId && options.sessionId === this.deps.session.getSessionId();
 
     const untrackSearch = isCurrentSession && this.deps.session.trackSearch({ abort });
 
-    return doPartialSearch<IEsSearchResponse>(
-      () => this.runSearch(request, { ...options, strategy, abortSignal: combinedSignal }),
-      (requestId) =>
-        this.runSearch(
-          { ...request, id: requestId },
-          { ...options, strategy, abortSignal: combinedSignal }
-        ),
-      (r) => !r.isRunning,
-      (response) => response.id,
-      id,
-      { pollInterval }
-    ).pipe(
-      tap((r) => {
-        id = r.id ?? id;
-      }),
-      throwOnEsError(),
-      takeUntil(from(abortedPromise.promise)),
+    return pollSearch(search, { ...options, abortSignal: combinedSignal }).pipe(
+      tap((response) => (id = response.id)),
       catchError((e: AbortError) => {
-        if (id) {
-          this.deps.http.delete(`/internal/search/${strategy}/${id}`);
-        }
-
-        return throwError(this.handleSearchError(e, request, timeoutSignal, options));
+        if (id) this.deps.http.delete(`/internal/search/${strategy}/${id}`);
+        return throwError(this.handleSearchError(e, timeoutSignal, options));
       }),
       finalize(() => {
         this.pendingCount$.next(this.pendingCount$.getValue() - 1);
         cleanup();
-        abortedPromise.cleanup();
         if (untrackSearch) {
           untrackSearch();
         }
